# api-docs/swagger.yaml
openapi: 3.0.0
info:
  title: Trae API
  description: API for Ticket Reservation and Event System
  version: 1.0.0
  contact:
    name: Trae Team
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/events:
    get:
      tags:
        - Events
      summary: List all events
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventResponse'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      tags:
        - Events
      summary: Create a new event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/events/{eventId}:
    get:
      tags:
        - Events
      summary: Get event by ID
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EventRequest:
      type: object
      required:
        - name
        - description
        - venue
        - date
        - price
        - total_tickets
      properties:
        name:
          type: string
        description:
          type: string
        venue:
          type: string
        date:
          type: string
          format: date-time
        price:
          type: number
          format: float
        total_tickets:
          type: integer
        image_url:
          type: string
          format: uri
    EventResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        venue:
          type: string
        date:
          type: string
          format: date-time
        price:
          type: number
          format: float
        total_tickets:
          type: integer
        available_tickets:
          type: integer
        image_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        status_code:
          type: integer